removed windows warnings

+++ src_tests/test10-SparseToolTiming.cc
@@ -88,7 +88,7 @@ test_CRow(
 ) {
 
   Utils::TicToc tm;
-  Vector<Real> res( v.size() );
+  Vector<Real> res( static_cast<integer>(v.size()) );
 
   tm.tic();
   REPEAT(N) res = crow * v;
@@ -127,7 +127,7 @@ test_CCol(
 ) {
 
   Utils::TicToc tm;
-  Vector<Real> res( v.size() );
+  Vector<Real> res( static_cast<integer>(v.size()) );
 
   tm.tic();
   REPEAT(N) res = ccol * v;
@@ -164,7 +164,7 @@ test_CCoor(
 ) {
 
   Utils::TicToc tm;
-  Vector<Real> res( v.size() );
+  Vector<Real> res( static_cast<integer>( v.size() ) );
 
   tm.tic();
   REPEAT(N) { res = ccoor * v; }
